Fixes unit tests
