+++ beagle/src/test/kotlin/br/com/zup/beagle/android/BaseSoLoaderTest.kt
@@ -16,14 +16,8 @@
 
 package br.com.zup.beagle.android
 
-import android.view.View
 import br.com.zup.beagle.android.setup.BeagleConfigurator
 import br.com.zup.beagle.android.setup.BeagleSdk
-import com.facebook.yoga.YogaNode
-import com.facebook.yoga.YogaNodeFactory
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.mockkStatic
 import org.junit.After
 import org.junit.Before
 
@@ -40,14 +34,5 @@ open class BaseSoLoaderTest : BaseTest() {
     @After
     fun teardown() {
         BeagleSdk.deinitForTest()
-    }
-
-    private fun mockYoga() {
-        val yogaNode = mockk<YogaNode>(relaxed = true, relaxUnitFun = true)
-        val view = View(application)
-        mockkStatic(YogaNode::class)
-        mockkStatic(YogaNodeFactory::class)
-        every { YogaNodeFactory.create() } returns yogaNode
-        every { yogaNode.data } returns view
     }
 }
