+++ beagle/src/test/kotlin/br/com/zup/beagle/android/BaseTest.kt
@@ -17,22 +17,23 @@
 package br.com.zup.beagle.android
 
 import android.app.Application
+import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.test.core.app.ApplicationProvider
-import br.com.zup.beagle.android.data.serializer.BeagleJsonSerializer
-import br.com.zup.beagle.android.data.serializer.BeagleMoshi
 import br.com.zup.beagle.android.setup.BeagleConfigurator
 import br.com.zup.beagle.android.setup.BeagleEnvironment
 import br.com.zup.beagle.android.setup.BeagleSdk
 import br.com.zup.beagle.android.widget.ActivityRootView
-import br.com.zup.beagle.android.widget.core.ServerDrivenComponent
+import com.facebook.yoga.YogaNode
+import com.facebook.yoga.YogaNodeFactory
 import com.squareup.moshi.Moshi
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkConstructor
 import io.mockk.mockkObject
+import io.mockk.mockkStatic
 import io.mockk.unmockkAll
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
@@ -52,9 +53,19 @@ abstract class BaseTest {
         mockBeagleEnvironment()
         every { rootView.activity } returns activity
         every { rootView.getScreenId() } returns "screen_id"
+        mockYoga()
         BeagleSdk.setInTestMode()
         MyBeagleSetup().init(application)
         beagleConfigurator = BeagleConfigurator.configurator
+    }
+
+    protected fun mockYoga() {
+        val yogaNode = mockk<YogaNode>(relaxed = true, relaxUnitFun = true)
+        val view = View(application)
+        mockkStatic(YogaNode::class)
+        mockkStatic(YogaNodeFactory::class)
+        every { YogaNodeFactory.create() } returns yogaNode
+        every { yogaNode.data } returns view
     }
 
     @AfterAll
